Refactored place page tests (#5113)
* Moved all place page tests into the shared_tests folder since the new
place pages are enabled on custom DC
* Removed all tests for the legacy place page (removed tests that use
disable_dev_places=true).
* Refactored a few of the legacy place page tests to run on the new
place page for better coverage

Ill remove the legacy place page entirely in a follow-up
