+++ server/webdriver/shared_tests/place_explorer_test.py
@@ -12,12 +12,18 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import pytest
+from selenium.webdriver.common.action_chains import ActionChains
 from selenium.webdriver.common.by import By
 from selenium.webdriver.support import expected_conditions as EC
 from selenium.webdriver.support.ui import WebDriverWait
 
+from server.routes.dev_place.utils import ORDERED_TOPICS
 from server.webdriver import shared
 from server.webdriver.base_utils import find_elem
+from server.webdriver.base_utils import find_elems
+from server.webdriver.base_utils import scroll_to_elem
+from server.webdriver.base_utils import wait_for_text
 
 MTV_URL = '/place/geoId/0649670'
 USA_URL = '/place/country/USA'
@@ -40,43 +46,35 @@ class PlaceExplorerTestMixin():
 
   def test_page_serve_usa(self):
     """Test the place explorer page for USA can be loaded successfullly."""
-    title_text = "United States of America - Place Explorer - " + self.dc_title_string
+    title_text = "United States of America - " + self.dc_title_string
     place_type_text = "Country in North America"
 
     # Load USA page.
-    self.driver.get(self.url_ + USA_URL + '?disable_dev_places=true')
-
-    # Assert 200 HTTP code: successful page load.
-    self.assertEqual(shared.safe_url_open(self.driver.current_url), 200)
-
-    # Assert 200 HTTP code: successful JS generation.
-    self.assertEqual(shared.safe_url_open(self.url_ + "/place.js"), 200)
+    self.driver.get(self.url_ + USA_URL)
 
     # Wait until the page loads and the title is correct.
     WebDriverWait(self.driver,
                   self.TIMEOUT_SEC).until(EC.title_contains(title_text))
     self.assertEqual(title_text, self.driver.title)
 
-    # Wait until the place type is correct.
-    element_present = EC.text_to_be_present_in_element((By.ID, 'place-type'),
-                                                       place_type_text)
-    WebDriverWait(self.driver, self.TIMEOUT_SEC).until(element_present)
+    # Wait for place name to load
+    place_name_present = EC.text_to_be_present_in_element(
+        (By.CSS_SELECTOR, '[data-testid="place-name"]'),
+        'United States of America')
+    WebDriverWait(self.driver, self.TIMEOUT_SEC).until(place_name_present)
 
-    # Assert place title is correct.
-    title = self.driver.find_element(By.ID, "place-name")
-    self.assertEqual("United States of America", title.text)
-
-    # Assert place type is correct.
-    subtitle = self.driver.find_element(By.ID, "place-type")
-    self.assertEqual("Country in North America", subtitle.text)
+    # Wait for subheader to load and contain "Country in"
+    subheader_present = EC.text_to_be_present_in_element(
+        (By.CLASS_NAME, 'subheader'), 'Country in')
+    WebDriverWait(self.driver, self.TIMEOUT_SEC).until(subheader_present)
 
   def test_page_serve_mtv(self):
     """Test the place explorer page for MTV can be loaded successfullly."""
-    place_type_title = "City in Santa Clara County, California, United States of America, North America"
-    title_text = "Mountain View - Place Explorer - " + self.dc_title_string
+    place_type_title = "City in Santa Clara County, California, United States of America, North America, World"
+    title_text = "Mountain View - " + self.dc_title_string
 
     # Load Mountain View Page.
-    self.driver.get(self.url_ + MTV_URL + '?disable_dev_places=true')
+    self.driver.get(self.url_ + MTV_URL)
 
     # Wait until the page loads and the title is correct.
     WebDriverWait(self.driver,
@@ -85,75 +83,67 @@ class PlaceExplorerTestMixin():
     # Assert title is correct.
     self.assertEqual(title_text, self.driver.title)
 
-    # Wait until the place name has loaded.
-    element_present = EC.text_to_be_present_in_element((By.ID, 'place-name'),
-                                                       "Mountain View")
-    WebDriverWait(self.driver, self.TIMEOUT_SEC).until(element_present)
-    place_name = self.driver.find_element(By.ID, "place-name").text
+    # Wait for place name to load
+    place_name_present = EC.text_to_be_present_in_element(
+        (By.CSS_SELECTOR, '[data-testid="place-name"]'), 'Mountain View')
+    WebDriverWait(self.driver, self.TIMEOUT_SEC).until(place_name_present)
 
-    # Assert place name is correct.
-    self.assertEqual("Mountain View", place_name)
+    # Wait for subheader to load and contain "City in"
+    subheader_present = EC.text_to_be_present_in_element(
+        (By.CLASS_NAME, 'subheader'), place_type_title)
+    WebDriverWait(self.driver, self.TIMEOUT_SEC).until(subheader_present)
 
-    # Wait until the place type has loaded.
-    element_present = EC.text_to_be_present_in_element((By.ID, 'place-type'),
-                                                       place_type_title)
-    WebDriverWait(self.driver, self.TIMEOUT_SEC).until(element_present)
+    # Wait for summary container to load
+    summary_title_present = EC.text_to_be_present_in_element(
+        (By.CSS_SELECTOR, '[data-testid="place-overview-summary-title"]'),
+        'Summary')
+    WebDriverWait(self.driver, self.TIMEOUT_SEC).until(summary_title_present)
 
-    # Assert place type is correct.
-    place_type = self.driver.find_element(By.ID, "place-type").text
-    self.assertEqual(place_type_title, place_type)
+    # Wait for overview table to load
+    table_present = EC.presence_of_element_located(
+        (By.CSS_SELECTOR, '[data-testid="place-overview-table"]'))
+    WebDriverWait(self.driver, self.TIMEOUT_SEC).until(table_present)
 
-    # Wait until place overview tile has loaded.
-    element_present = EC.presence_of_element_located(
-        (By.CLASS_NAME, 'overview-tile'))
-    WebDriverWait(self.driver, self.TIMEOUT_SEC).until(element_present)
-
-    # Wait until the population highlight has loaded.
-    element_present = EC.presence_of_element_located(
-        (By.ID, 'place-highlight-in-overview'))
-    WebDriverWait(self.driver, self.TIMEOUT_SEC).until(element_present)
-
-    # Assert population highlight is correct.
-    place_type = self.driver.find_element(By.ID,
-                                          "place-highlight-in-overview").text
-    self.assertTrue(place_type.startswith("Population:"))
+    # Check that unemployment rate is in the table
+    table_text = self.driver.find_element(
+        By.CSS_SELECTOR, '[data-testid="place-overview-table"]').text
+    self.assertIn("Unemployment Rate", table_text)
 
   def test_demographics_link(self):
     """Test the demographics link can work correctly."""
     title_text = "Median age by gender: states near California"
     # Load California page.
-    self.driver.get(self.url_ + CA_URL + '?disable_dev_places=true')
+    self.driver.get(self.url_ + CA_URL)
 
-    # Wait until the Demographics link is present.
-    element_present = EC.presence_of_element_located(
-        (By.XPATH, '//*[@id="Demographics"]/a'))
-    WebDriverWait(self.driver, self.TIMEOUT_SEC).until(element_present)
+    # Find demographic link in explore topics box
+    topics_for_ca = [
+        "Economics", "Health", "Equity", "Crime", "Education", "Demographics",
+        "Housing", "Environment", "Energy"
+    ]
+    shared.assert_topics(self,
+                         self.driver,
+                         path_to_topics=['explore-topics-box'],
+                         classname='item-list-item',
+                         expected_topics=topics_for_ca)
 
-    # Find and click on the Demographics URL.
-    demographics = self.driver.find_element(By.XPATH,
-                                            '//*[@id="Demographics"]/a')
+    demographics = find_elem(
+        self.driver,
+        by=By.CSS_SELECTOR,
+        value='.item-list-item[data-testid="Demographics"] a')
     demographics.click()
-    self.driver.get(self.driver.current_url + '&disable_dev_places=true')
+    self.driver.get(self.driver.current_url)
 
-    # Wait until the new page has loaded.
-    element_present = EC.presence_of_element_located(
-        (By.XPATH, '//*[@id="main-pane"]/section[4]/div/div[2]/div/h4'))
-    WebDriverWait(self.driver, self.TIMEOUT_SEC).until(element_present)
+    # Wait until the new page has loaded and check for text
+    wait_for_text(self.driver, "Demographics", By.CSS_SELECTOR,
+                  'span[data-testid="place-name"]')
 
-    # Assert that Demographics is part of the new url.
+    # Assert that Demographics is part of the new url
     self.assertTrue("Demographics" in self.driver.current_url)
-
-    # Get the chart_title text.
-    chart_title = self.driver.find_element(
-        By.XPATH, '//*[@id="main-pane"]/section[4]/div/div[2]/div/h4').text
-
-    # Assert chart title is correct.
-    self.assertIn(title_text, chart_title)
 
   def test_demographics_redirect_link(self):
     """Test a place page with demographics after a redirect."""
     # Load California's Demographics page.
-    start_url = self.url_ + '/place?dcid=geoId/06&category=Demographics&utm_medium=um&disable_dev_places=true'
+    start_url = self.url_ + '/place?dcid=geoId/06&category=Demographics&utm_medium=um'
     self.driver.get(start_url)
 
     # Wait for redirect.
@@ -163,15 +153,9 @@ class PlaceExplorerTestMixin():
     self.assertTrue("category=Demographics" in self.driver.current_url)
     self.assertTrue("utm_medium=um" in self.driver.current_url)
 
-    element_present = EC.presence_of_element_located(
-        (By.XPATH, '//*[@id="main-pane"]/section[4]/div/div[2]/div/h4'))
-    WebDriverWait(self.driver, self.TIMEOUT_SEC * 2).until(element_present)
-
-    chart_title = self.driver.find_element(
-        By.XPATH, '//*[@id="main-pane"]/section[4]/div/div[2]/div/h4').text
-
-    # Assert chart title is correct.
-    self.assertIn("Median age by gender: states near California", chart_title)
+    # Wait for place name to show correct text
+    wait_for_text(self.driver, "California • Demographics", By.CSS_SELECTOR,
+                  '[data-testid="place-name"]')
 
   def test_explorer_redirect_place_explorer(self):
     """Test the redirection from explore to place explore for single place queries"""
@@ -195,137 +179,6 @@ class PlaceExplorerTestMixin():
     # Assert page title is correct.
     WebDriverWait(self.driver, self.TIMEOUT_SEC).until(
         EC.title_contains('United States of America'))
-
-  def test_ranking_chart_present(self):
-    """Test basic ranking chart."""
-    CHART_TITLE_MUST_INCLUDE = "rankings"
-    CHART_SUBTITLE_MUST_INCLUDE = "California ranks"
-    CHART_UNIT_HIGHEST_TITLE = "Highest"
-    CHART_UNIT_LOWEST_TITLE = "Lowest"
-    ca_economics_url = CA_URL + "?category=Economics"
-    first_rank = "1."
-    last_rank = "52."
-
-    # Load California's economincs page.
-    self.driver.get(self.url_ + ca_economics_url + '&disable_dev_places=true')
-
-    # Wait for the presense of the chart title element.
-    chart_title_present = EC.presence_of_element_located(
-        (By.XPATH, '//*[@id="main-pane"]/section[5]/div/div[6]/div/h4'))
-    WebDriverWait(self.driver, self.TIMEOUT_SEC).until(chart_title_present)
-
-    # Check the chart title text.
-    chart_title = self.driver.find_element(
-        By.XPATH, '//*[@id="main-pane"]/section[5]/div/div[6]/div/h4').text
-    self.assertTrue = (CHART_TITLE_MUST_INCLUDE in chart_title)
-
-    # Wait for the presence of the chart subtitle element.
-    chart_subtitle_present = EC.presence_of_element_located(
-        (By.XPATH, '//*[@id="main-pane"]/section[5]/div/div[6]/div/div/div/h4'))
-    WebDriverWait(self.driver, self.TIMEOUT_SEC).until(chart_subtitle_present)
-
-    # Check the chart subtitle text.
-    chart_subtitle = self.driver.find_element(
-        By.XPATH,
-        '//*[@id="main-pane"]/section[5]/div/div[6]/div/div/div/h4').text
-    self.assertTrue = (CHART_SUBTITLE_MUST_INCLUDE in chart_subtitle)
-
-    # Wait for the presence of the chart_unit_highest and chart_unit_lowest element.
-    chart_unit_highest = EC.presence_of_element_located(
-        (By.XPATH,
-         '//*[@id="main-pane"]/section[5]/div/div[6]/div/div/div/div/div[1]'))
-    WebDriverWait(self.driver, self.TIMEOUT_SEC).until(chart_unit_highest)
-    chart_unit_lowest = EC.presence_of_element_located(
-        (By.XPATH,
-         '//*[@id="main-pane"]/section[5]/div/div[6]/div/div/div/div/div[2]'))
-    WebDriverWait(self.driver, self.TIMEOUT_SEC).until(chart_unit_lowest)
-
-    # Check the title text.
-    chart_unit_highest_title = self.driver.find_element(
-        By.XPATH,
-        '//*[@id="main-pane"]/section[5]/div/div[6]/div/div/div/div/div[1]/div/h4'
-    ).text
-    self.assertTrue = (CHART_UNIT_HIGHEST_TITLE in chart_unit_highest_title)
-
-    chart_unit_lowest_title = self.driver.find_element(
-        By.XPATH,
-        '//*[@id="main-pane"]/section[5]/div/div[6]/div/div/div/div/div[2]/div/h4'
-    ).text
-    self.assertTrue = (CHART_UNIT_LOWEST_TITLE in chart_unit_lowest_title)
-
-    # Wait for the presence of the row.
-    chart_unit_highest_row = EC.presence_of_element_located((
-        By.XPATH,
-        '//*[@id="main-pane"]/section[5]/div/div[6]/div/div/div/div/div[1]/table/tbody/tr[1]/td'
-    ))
-    WebDriverWait(self.driver, self.TIMEOUT_SEC).until(chart_unit_highest_row)
-    chart_unit_lowest_row = EC.presence_of_element_located((
-        By.XPATH,
-        '//*[@id="main-pane"]/section[5]/div/div[6]/div/div/div/div/div[2]/table/tbody/tr[1]/td'
-    ))
-    WebDriverWait(self.driver, self.TIMEOUT_SEC).until(chart_unit_lowest_row)
-
-    # Check the rank.
-    chart_unit_highest_row_rank = self.driver.find_elements(
-        By.XPATH,
-        '//*[@id="main-pane"]/section[5]/div/div[6]/div/div/div/div/div[1]/table/tbody/tr[1]/td'
-    )
-    self.assertEqual(chart_unit_highest_row_rank[0].text, first_rank)
-
-    chart_unit_lowest_row_rank = self.driver.find_elements(
-        By.XPATH,
-        '//*[@id="main-pane"]/section[5]/div/div[6]/div/div/div/div/div[2]/table/tbody/tr[1]/td'
-    )
-    self.assertEqual(chart_unit_lowest_row_rank[0].text, last_rank)
-
-  def test_ranking_chart_redirect_link(self):
-    """Test the redirect link can work correctly."""
-    ca_economics_url = CA_URL + "?category=Economics&disable_dev_places=true"
-    self.driver.get(self.url_ + ca_economics_url)
-
-    xpath = '//*[@id="main-pane"]/section[5]/div/div[6]/div/div/div/div/div[1]/table/tbody/tr[1]/td[2]/a'
-    # Wait for the presence of place name.
-    WebDriverWait(self.driver, self.TIMEOUT_SEC).until(
-        lambda driver: element_has_text(driver, (By.XPATH, xpath)))
-
-    # Click the place name.
-    place_name = self.driver.find_element(By.XPATH, xpath)
-    place_name_text = place_name.text
-    place_name.click()
-    shared.wait_for_loading(self.driver)
-    self.driver.get(self.driver.current_url + '?disable_dev_places=true')
-    shared.wait_for_loading(self.driver)
-
-    # Wait for the presence of new page title.
-    title_present = EC.presence_of_element_located((By.ID, 'place-name'))
-    WebDriverWait(self.driver, self.TIMEOUT_SEC).until(title_present)
-
-    # Check the title text
-    page_title = self.driver.find_element(By.ID, 'place-name').text
-    self.assertEqual(page_title, place_name_text)
-
-  def test_export_chart_data(self):
-    """Tests the export chart data button works correctly for group bar charts."""
-    # Load CA housing page
-    ca_housing_url = CA_URL + "?category=Housing&disable_dev_places=true"
-    self.driver.get(self.url_ + ca_housing_url)
-
-    # Wait for trend chart to load
-    trend_chart = EC.presence_of_element_located(
-        (By.CSS_SELECTOR, '[data-testclass~="is-snapshot"]'))
-    WebDriverWait(self.driver, self.TIMEOUT_SEC).until(trend_chart)
-
-    # Find and click export link for group bar chart
-    export_link = self.driver.find_element(
-        By.XPATH,
-        "//*[@data-testclass='is-snapshot chart-type-GROUP_BAR']//div[contains(@class,'outlinks')]//a[text()='Export']"
-    )
-    export_link.click()
-
-    # Wait for entity DCID text to appear in dialog
-    entity_dcid_present = EC.text_to_be_present_in_element(
-        (By.CLASS_NAME, "copy-svg"), "Entity DCID")
-    WebDriverWait(self.driver, self.TIMEOUT_SEC).until(entity_dcid_present)
 
   def test_canonical_links_in_html_head(self):
     """Test that canonical and alternate language links are present in HTML head"""
@@ -382,14 +235,11 @@ class PlaceExplorerTestMixin():
     self.driver.get(self.url_ + NO_DATA_NEIGHBORHOOD_URL)
 
     # Wait for subheader to load and contain "Neighborhood in"
-    subheader_present = EC.text_to_be_present_in_element(
-        (By.CLASS_NAME, 'subheader'), 'Neighborhood in')
-    WebDriverWait(self.driver, self.TIMEOUT_SEC).until(subheader_present)
+    wait_for_text(self.driver, "Neighborhood in", By.CLASS_NAME, 'subheader')
 
     # Wait for no data message to appear
-    no_data_present = EC.text_to_be_present_in_element(
-        (By.CLASS_NAME, 'page-content-container'), 'No data found for')
-    WebDriverWait(self.driver, self.TIMEOUT_SEC).until(no_data_present)
+    wait_for_text(self.driver, "No data found for", By.CLASS_NAME,
+                  'page-content-container')
 
   def test_place_category_placeholder(self):
     """Test that the place category placeholder is correct."""
@@ -398,13 +248,393 @@ class PlaceExplorerTestMixin():
     self.driver.get(self.url_ + ca_housing_url)
 
     # Wait for search input to be present
-    search_input_present = EC.presence_of_element_located(
-        (By.ID, "query-search-input"))
-    WebDriverWait(self.driver, self.TIMEOUT_SEC).until(search_input_present)
+    search_input = find_elem(self.driver, by=By.ID, value="query-search-input")
 
     # Wait for placeholder text to update to expected value
-    def placeholder_has_text(driver):
-      element = driver.find_element(By.ID, "query-search-input")
-      return element.get_attribute("placeholder") == "Housing in California"
+    WebDriverWait(
+        self.driver,
+        self.TIMEOUT_SEC).until(lambda driver: search_input.get_attribute(
+            "placeholder") == "Housing in California")
 
-    WebDriverWait(self.driver, self.TIMEOUT_SEC).until(placeholder_has_text)
+  def test_dev_place_overview_world(self):
+    """Ensure place page revamp World page works"""
+    self.driver.get(self.url_ + '/place/Earth')
+
+    # Assert the subheader contains the parent places.
+    self.assertIsNotNone(find_elem(self.driver, value='place-info'))
+    self.assertIsNone(find_elem(self.driver, value='subheader'))
+
+    # Asert the related places box exists
+    self.assertEqual(
+        find_elem(self.driver, value='related-places-callout').text,
+        'Places in World')
+
+    # Assert we have data for all expected topics.
+    topics_for_world = [
+        "Economics", "Health", "Equity", "Crime", "Demographics", "Housing",
+        "Environment", "Energy"
+    ]
+    shared.assert_topics(self,
+                         self.driver,
+                         path_to_topics=['explore-topics-box'],
+                         classname='item-list-item',
+                         expected_topics=topics_for_world)
+
+    # And that the categories have data in the overview
+    block_titles = find_elems(self.driver, value='block-title-text')
+    self.assertEqual(set([block.text for block in block_titles]),
+                     set(topics_for_world))
+
+    # Assert that every category is expected, and has at least one chart
+    category_containers = find_elems(self.driver,
+                                     value='category',
+                                     path_to_elem=['charts-container'])
+    self.assertEqual(len(category_containers), len(topics_for_world))
+    for category_container in category_containers:
+      chart_containers = find_elems(category_container, value='chart-container')
+      self.assertGreater(len(chart_containers), 0)
+
+  def test_dev_place_demographics_world(self):
+    """Ensure place page revamp World page works"""
+    self.driver.get(self.url_ + '/place/Earth?category=Demographics')
+
+    # Assert we have data for all expected topics.
+    topics_for_world = [
+        "Economics", "Health", "Equity", "Crime", "Demographics", "Housing",
+        "Environment", "Energy"
+    ]
+    shared.assert_topics(self,
+                         self.driver,
+                         path_to_topics=['explore-topics-box'],
+                         classname='item-list-item',
+                         expected_topics=topics_for_world)
+
+    # Scroll to a map chart.
+    map_container = scroll_to_elem(self.driver, value='map-chart')
+    self.assertIsNotNone(map_container)
+
+    map_geo_regions = find_elem(map_container,
+                                by=By.ID,
+                                value='map-geo-regions',
+                                path_to_elem=['map-items'])
+    self.assertIsNotNone(map_geo_regions)
+    self.assertEqual(
+        len(find_elems(map_geo_regions, by=By.TAG_NAME, value='path')),
+        238)  # country count.
+
+  def test_dev_place_overview_world_redirect(self):
+    """Ensure click on place link from the demographics page redirects to the overview page"""
+    self.driver.get(self.url_ + '/place/Earth?category=Demographics')
+
+    # Click the place link and verify it takes us to the overview page
+    place_link = find_elem(self.driver,
+                           by=By.CSS_SELECTOR,
+                           value='.place-info-link')
+    place_link.click()
+
+    # Wait for URL change and page load
+    WebDriverWait(self.driver, self.TIMEOUT_SEC).until(
+        EC.url_to_be(self.url_ + '/place/Earth'))
+
+  def test_dev_place_overview_california(self):
+    """Ensure experimental dev place page content loads"""
+    self.driver.get(self.url_ + '/place/geoId/06')
+
+    # Assert the subheader contains the parent places.
+    self.assertIsNotNone(find_elem(self.driver, value='place-info'))
+    self.assertEqual(
+        find_elem(self.driver, value='subheader').text,
+        'State in United States of America, North America, World')
+
+    # Asert the related places box exists
+    self.assertEqual(
+        find_elem(self.driver, value='related-places-callout').text,
+        'Places in California')
+
+    # Assert the overview exists, has a summary and a map.
+    self.assertNotEqual(
+        len(
+            find_elem(self.driver, by=By.CSS_SELECTOR,
+                      value='.place-summary').text), "")
+    self.assertIsNotNone(find_elem(self.driver, value='map-container'))
+
+    # Assert the key demographics table has data
+    self.assertEqual(
+        len(
+            find_elems(self.driver,
+                       value='key-demographics-row',
+                       path_to_elem=['key-demographics-table'])), 4)
+
+    shared.assert_topics(self,
+                         self.driver,
+                         path_to_topics=['explore-topics-box'],
+                         classname='item-list-item',
+                         expected_topics=ORDERED_TOPICS)
+
+    # And that the categories have data in the overview
+    topics_in_overview = [
+        "Economics", "Health", "Equity", "Crime", "Education", "Demographics",
+        "Housing", "Energy"
+    ]
+    block_titles = find_elems(self.driver, value='block-title-text')
+    self.assertEqual([block.text for block in block_titles], topics_in_overview)
+
+    # Assert that every category is expected, and has at least one chart
+    category_containers = find_elems(self.driver,
+                                     value='category',
+                                     path_to_elem=['charts-container'])
+    self.assertEqual(len(category_containers), len(topics_in_overview))
+    for category_container in category_containers:
+      chart_containers = find_elems(category_container, value='chart-container')
+      self.assertGreater(len(chart_containers), 0)
+
+  def test_dev_place_overview_san_mateo_county(self):
+    """Ensure experimental dev place page content loads"""
+    self.driver.get(self.url_ + '/place/geoId/06081')
+
+    # Assert the subheader contains the parent places.
+    self.assertIsNotNone(find_elem(self.driver, value='place-info'))
+    self.assertEqual(
+        find_elem(self.driver, value='subheader').text,
+        'County in California, United States of America, North America, World')
+
+    # Asert the related places box exists
+    self.assertEqual(
+        find_elem(self.driver, value='related-places-callout').text,
+        'Places in San Mateo County')
+
+    # Assert the overview exists, has a summary and a map.
+    self.assertNotEqual(
+        len(
+            find_elem(self.driver, by=By.CSS_SELECTOR,
+                      value='.place-summary').text), "")
+    self.assertIsNotNone(find_elem(self.driver, value='map-container'))
+
+    # Assert the key demographics table has data
+    self.assertEqual(
+        len(
+            find_elems(self.driver,
+                       value='key-demographics-row',
+                       path_to_elem=['key-demographics-table'])), 4)
+
+    # And that the categories have data
+    topics_in_overview = [
+        "Economics", "Health", "Equity", "Crime", "Education", "Demographics",
+        "Housing"
+    ]
+    topics_with_data = topics_in_overview + ["Environment"]
+
+    shared.assert_topics(self,
+                         self.driver,
+                         path_to_topics=['explore-topics-box'],
+                         classname='item-list-item',
+                         expected_topics=topics_with_data)
+
+    block_titles = find_elems(self.driver, value='block-title-text')
+    self.assertEqual([block.text for block in block_titles], topics_in_overview)
+
+    # Assert that every category is expected, and has at least one chart
+    category_containers = find_elems(self.driver,
+                                     value='category',
+                                     path_to_elem=['charts-container'])
+    self.assertEqual(len(category_containers), len(topics_in_overview))
+    for category_container in category_containers:
+      chart_containers = find_elems(category_container, value='chart-container')
+      self.assertGreater(len(chart_containers), 0)
+
+  def test_dev_place_overview_los_angeles(self):
+    """Ensure experimental dev place page content loads"""
+    self.driver.get(self.url_ + '/place/geoId/0644000')
+
+    # Assert the subheader contains the parent places.
+    self.assertIsNotNone(find_elem(self.driver, value='place-info'))
+    self.assertEqual(
+        find_elem(self.driver, value='subheader').text,
+        'City in Los Angeles County, California, United States of America, North America, World'
+    )
+
+    # Asert the related places box exists
+    self.assertEqual(
+        find_elem(self.driver, value='related-places-callout').text,
+        'Places in Los Angeles')
+
+    # Assert the overview exists, has a summary and a map.
+    self.assertNotEqual(
+        len(
+            find_elem(self.driver, by=By.CSS_SELECTOR,
+                      value='.place-summary').text), "")
+    self.assertIsNotNone(find_elem(self.driver, value='map-container'))
+
+    # Assert the key demographics table has data
+    self.assertEqual(
+        len(
+            find_elems(self.driver,
+                       value='key-demographics-row',
+                       path_to_elem=['key-demographics-table'])), 4)
+
+    # And that the categories have data
+    topics_in_overview = [
+        "Economics", "Health", "Equity", "Crime", "Education", "Demographics",
+        "Housing"
+    ]
+    topics_with_data = topics_in_overview + ["Environment"]
+
+    shared.assert_topics(self,
+                         self.driver,
+                         path_to_topics=['explore-topics-box'],
+                         classname='item-list-item',
+                         expected_topics=topics_with_data)
+
+    block_titles = find_elems(self.driver, value='block-title-text')
+    self.assertEqual([block.text for block in block_titles], topics_in_overview)
+
+    # Assert that every category is expected, and has at least one chart
+    category_containers = find_elems(self.driver,
+                                     value='category',
+                                     path_to_elem=['charts-container'])
+    self.assertEqual(len(category_containers), len(topics_in_overview))
+    for category_container in category_containers:
+      chart_containers = find_elems(category_container, value='chart-container')
+      self.assertGreater(len(chart_containers), 0)
+
+  def test_dev_place_overview_zip_90003(self):
+    """Ensure experimental dev place page content loads"""
+    self.driver.get(self.url_ + '/place/zip/90003')
+
+    # Assert the subheader contains the parent places.
+    self.assertIsNotNone(find_elem(self.driver, value='place-info'))
+    self.assertEqual(
+        find_elem(self.driver, value='subheader').text,
+        'ZIP Code in Los Angeles, Los Angeles County, California, United States of America, North America, World'
+    )
+
+    # Asert the related places box does not exist
+    self.assertIsNone(find_elem(self.driver, value='related-places-callout'))
+
+    # Assert the overview exists, has a map. Zips have no summaries.
+    self.assertIsNotNone(find_elem(self.driver, value='map-container'))
+
+    # Assert the key demographics table has data
+    self.assertEqual(
+        len(
+            find_elems(self.driver,
+                       value='key-demographics-row',
+                       path_to_elem=['key-demographics-table'])), 3)
+
+    # And that the categories have data
+    topics_in_overview = [
+        "Economics", "Health", "Equity", "Education", "Demographics", "Housing"
+    ]
+
+    shared.assert_topics(self,
+                         self.driver,
+                         path_to_topics=['explore-topics-box'],
+                         classname='item-list-item',
+                         expected_topics=topics_in_overview)
+
+    block_titles = find_elems(self.driver, value='block-title-text')
+    self.assertEqual([block.text for block in block_titles], topics_in_overview)
+
+    # Assert that every category is expected, and has at least one chart
+    category_containers = find_elems(self.driver,
+                                     value='category',
+                                     path_to_elem=['charts-container'])
+    self.assertEqual(len(category_containers), len(topics_in_overview))
+    for category_container in category_containers:
+      chart_containers = find_elems(category_container, value='chart-container')
+      self.assertGreater(len(chart_containers), 0)
+
+  def test_dev_place_overview_africa(self):
+    """Ensure experimental dev place page content loads data for a continent."""
+    self.driver.get(self.url_ + '/place/africa')
+
+    # Assert the subheader contains the parent places.
+    self.assertIsNotNone(find_elem(self.driver, value='place-info'))
+    self.assertEqual(
+        find_elem(self.driver, value='subheader').text, 'Continent in World')
+
+    # Asert the related places box exists
+    self.assertEqual(
+        find_elem(self.driver, value='related-places-callout').text,
+        'Places in Africa')
+
+    topics_for_africa = [
+        "Economics", "Health", "Equity", "Demographics", "Environment", "Energy"
+    ]
+    shared.assert_topics(self,
+                         self.driver,
+                         path_to_topics=['explore-topics-box'],
+                         classname='item-list-item',
+                         expected_topics=topics_for_africa)
+
+    # And that the categories have data in the overview
+    block_titles = find_elems(self.driver, value='block-title-text')
+    self.assertEqual([block.text for block in block_titles], topics_for_africa)
+
+    # Assert that every category is expected, and has at least one chart
+    category_containers = find_elems(self.driver,
+                                     value='category',
+                                     path_to_elem=['charts-container'])
+    self.assertEqual(len(category_containers), len(topics_for_africa))
+    for category_container in category_containers:
+      chart_containers = find_elems(category_container, value='chart-container')
+      self.assertGreater(len(chart_containers), 0)
+
+  def test_dev_place_chart_settings(self):
+    """Ensure the charts in the new place page contain the expected settings"""
+    self.driver.get(self.url_ + '/place/geoId/06')
+
+    self.assertTrue(
+        find_elem(self.driver,
+                  value='download-outlink',
+                  path_to_elem=['charts-container']).text, 'Download')
+
+    self.assertTrue(
+        find_elem(self.driver,
+                  value='explore-in-outlink',
+                  path_to_elem=['charts-container']).text,
+        'Explore in Timeline tool',
+    )
+
+  def test_explorer_redirect_place_explorer_populates_search_bar(self):
+    """Test the redirection from explore to place explore for single place queries populates the search bar from the URL query"""
+    usa_explore = '/explore#q=United%20States%20Of%20America'
+
+    start_url = self.url_ + usa_explore
+    self.driver.get(start_url)
+
+    # Assert 200 HTTP code: successful page load.
+    self.assertEqual(shared.safe_url_open(self.driver.current_url), 200)
+
+    # Wait for redirect and page load
+    redirect_finished = EC.url_changes(start_url)
+    WebDriverWait(self.driver, self.TIMEOUT_SEC).until(redirect_finished)
+    shared.wait_for_loading(self.driver)
+
+    # Ensure the query string is set in the NL Search Bar.
+    self.assertEqual(
+        find_elem(self.driver, by=By.ID,
+                  value='query-search-input').get_attribute('value'),
+        'United States Of America')
+
+  @pytest.mark.skip(reason="Fix theme compile error before re-enabling")
+  def test_dev_place_ai_spark_icon_hover(self):
+    self.driver.get(self.url_ + '/place/geoId/04')
+
+    # Find the icon element that triggers the tooltip on hover.
+    icon_element = find_elem(self.driver, value='spark-info-tooltip-icon')
+    self.assertIsNotNone(icon_element)
+
+    # Move the mouse to the icon element to trigger the hover.
+    actions = ActionChains(self.driver)
+    actions.move_to_element(icon_element).perform()
+
+    # Wait for the tooltip to become visible.
+    self.driver.implicitly_wait(2)
+
+    # Find the tooltip element.
+    tooltip_element = find_elem(self.driver, value='info-tooltip-hover')
+
+    # Assert that the tooltip element is visible.
+    self.assertTrue(tooltip_element.is_displayed())
+    self.assertIn("We use AI to summarize", tooltip_element.text)
