+++ server/webdriver/shared_tests/place_landing_test.py
@@ -98,77 +98,3 @@ class PlaceLandingTestMixin():
                                           '//*[@id="place-autocomplete"]')
     self.assertEqual(map_search.get_attribute('placeholder'),
                      'Укажите страну, штат, округ или город')
-
-  def test_place_landing_explore_more(self):
-    """Test place landing explore more link."""
-
-    self.driver.get(
-        self.url_ +
-        '/place/geoId/1714000?category=Education&disable_dev_places=true')
-
-    # Wait until the chart has loaded.
-    element_present = EC.presence_of_element_located(
-        (By.CLASS_NAME, 'chart-container'))
-    WebDriverWait(self.driver, self.TIMEOUT_SEC).until(element_present)
-
-    explore_more = self.driver.find_element(
-        By.XPATH,
-        '//*[@id="main-pane"]/section[1]/div/div[2]/div/footer/div[2]/a[2]')
-    self.assertEqual(explore_more.text, 'Explore More ›')
-
-    explore_more.click()
-
-    # Wait for the new page to open in a new tab
-    new_page_opened = EC.number_of_windows_to_be(2)
-    WebDriverWait(self.driver, self.TIMEOUT_SEC).until(new_page_opened)
-
-    # Switch tabs to the page for the timeline tool
-    new_page = self.driver.window_handles[-1]
-    self.driver.switch_to.window(new_page)
-
-    # Assert timelines page loaded
-    new_page_title = 'Timelines Explorer - ' + self.dc_title_string
-    WebDriverWait(self.driver,
-                  self.TIMEOUT_SEC).until(EC.title_contains(new_page_title))
-
-    # Wait until the group of charts has loaded.
-    element_present = EC.presence_of_element_located((By.ID, 'chart-region'))
-    WebDriverWait(self.driver, self.TIMEOUT_SEC).until(element_present)
-
-    # Store a list of all the charts.
-    chart_region = self.driver.find_element(By.XPATH, '//*[@id="chart-region"]')
-    charts = chart_region.find_elements(By.CLASS_NAME, 'card')
-    # Assert there are three charts.
-    self.assertEqual(len(charts), 1)
-    # Wait until the charts are drawn.
-    element_present = EC.presence_of_element_located(
-        (By.CLASS_NAME, 'legend-text'))
-    WebDriverWait(self.driver, self.TIMEOUT_SEC).until(element_present)
-    # Assert first chart has 36 lines (ie. has data)
-    chart_lines = charts[0].find_elements(By.CLASS_NAME, 'line')
-    self.assertGreater(len(chart_lines), 10)
-
-  def test_place_landing_url_param(self):
-    """Test place landing url parameters."""
-
-    # Test for ?category=Education
-    self.driver.get(
-        self.url_ +
-        '/place/geoId/1714000?category=Education&disable_dev_places=true')
-    element_present = EC.presence_of_element_located(
-        (By.CLASS_NAME, 'chart-container'))
-    WebDriverWait(self.driver, self.TIMEOUT_SEC).until(element_present)
-    link = self.driver.find_element(
-        By.XPATH, '//*[@id="nav-topics"]/li[6]/ul/div/li[1]/a')
-    self.assertEqual(link.text, 'Education attainment')
-
-    # Test for ?topic=Education
-    self.driver.get(
-        self.url_ +
-        '/place/geoId/1714000?topic=Education&disable_dev_places=true')
-    element_present = EC.presence_of_element_located(
-        (By.CLASS_NAME, 'chart-container'))
-    WebDriverWait(self.driver, self.TIMEOUT_SEC).until(element_present)
-    link = self.driver.find_element(
-        By.XPATH, '//*[@id="nav-topics"]/li[6]/ul/div/li[1]/a')
-    self.assertEqual(link.text, 'Education attainment')
